Move parser package into parser/ subdirectory
Relocate the root-level parser files into parser/ so the import path
becomes github.com/sqlc-dev/marino/parser instead of the awkward
github.com/sqlc-dev/marino (where the package name didn't match the
import's last segment). Subpackages (ast, mysql, types, etc.) remain
at the module root. Also remove the leftover FOO.txt placeholder.
