+++ parser/consistent_test.go
