+++ parser/hintparser_test.go
@@ -16,7 +16,7 @@ package parser_test
 import (
 	"testing"
 
-	"github.com/sqlc-dev/marino"
+	"github.com/sqlc-dev/marino/parser"
 	"github.com/sqlc-dev/marino/ast"
 	"github.com/sqlc-dev/marino/mysql"
 	"github.com/stretchr/testify/require"
