+++ parser/parser_test.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 
 	"github.com/pingcap/errors"
-	"github.com/sqlc-dev/marino"
+	"github.com/sqlc-dev/marino/parser"
 	"github.com/sqlc-dev/marino/ast"
 	"github.com/sqlc-dev/marino/charset"
 	. "github.com/sqlc-dev/marino/format"
