+++ types/field_type_test.go
@@ -17,7 +17,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/sqlc-dev/marino"
+	"github.com/sqlc-dev/marino/parser"
 	"github.com/sqlc-dev/marino/ast"
 	"github.com/sqlc-dev/marino/charset"
 	"github.com/sqlc-dev/marino/mysql"
